codal: WIP refactoring timer functionality
The simplicity of receiving a regular callback previously in codal was really beneficial. There was not much overhead compared to having a listener listening for specific timer events.

This commit begins a series of commits to restore and build on this functionality.

A new static member of DeviceComponent holds pointers to all of the DeviceComponents created a runtime. These components now add themselves to an array pending a future callback implementation.

To receive a regular callback, a Component now simply toggles a bit in the status field.

The status field is now a uint16_t, where "system" flags are implemented from the MSB to the LSB, and user flags begin from the LSB to the MSB.