codal: refactored Timer interface, added periodic and idle callbacks
The timer interface now accepts an id, so that custom, unique events can be conf
igured, and subsequently listened too.

Timer.cancel(id, value) has also been added, so that ClockEvents that were previ
ously configured can be removed.

DeviceComponent now has an array, that is added to on construction. A timer event and listener are also configured. When an event is fired, components are stepped through, and periodicCallback is invoked if requested via setting the DEVICE_COMPONENT_STATUS_SYSTEM_TICK bit.