DataStream now always pulls a buffer on a synchronous pullRequest()
Previous revision of DataStream would block the calling fiber when the buffer
was full, then subsequenxtly pull() a puffer when space became available.

This approach can lead o complexity, as the upstream component will be left with
a valid buffer in its output, but no record of this buffer being in transit.

This patch causes the upstream buffer to be pulled into a local reference
before the fiber is blocked, effectively decoupling the upstream output
buffer.