Synthesizer::setFrequency(f, x) is now synchronous.
A call to setFrequency() with a playoutTime parameter now blocks the calling
fiber until the generated sample data has been added to the output buffer.
This makes creating sequences of tones much simpler.

Semantics of a call to setFrequency() with an undefined playoutTime is
unchanged - an asynchronous, unterminating playout of the given frequency.