Reduce fragmentation of buffers in Synthesizer
When in synchronous mode, the Synthesizer no longer creates a pullRequest()
for partially full buffers at the end of each requested tone - instead, the
space in the buffer carries over to the next tone.

Added a background idle task to flush the buffer if the user transparently
stops using sycnhronpus mode.