Remove erroneous descheduling of fiber in fiber_wake_on_event()
Previous versions would descheule the calling fiber before placing the fiber
of the wait queue, ready to receive events. The fiber would only be added to
the waitqueue the next time it was scheduled. This leaves a wide, and
undecessary, critical section.

Fibers are now added directly to the wait queue.