Added sensible defaults for APP source and output
with the future removal of codal.json and generation of codal.json by build.py, it is sensible to add defaults for application source and application output so that the size of codal.json is small and can be easily specified.

In this commit I also changed the location of the example application for a target so that an arbitrary number of files can be copied into the source tree of a user.