build: Add an option to checkout a specific revision of the target
This patch adds the option -r/--revision to specify a specific revision
of the target. The revision specified can be a tag or a branch.

The idea is to allow a simpler way to checkout a feature branch for the
target from which the other libraries can be udated based on the
target.json file instead of having to checkout manually all of the
libraries.

Signed-off-by: Raphael Gault <raphael.gault@arm.com>